Re-organized the awesome folder
Much more like the original awesome directory structure, though I still
need to move all the lua/ code out.
