+++ awesome/src/common/class.rs
