+++ awesome/src/common/mod.rs
@@ -1,0 +1,5 @@
+pub mod class;
+pub mod object;
+pub mod property;
+pub mod signal;
+pub mod xproperty;
