+++ awesome/src/common/object.rs
